+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Allow for generic `prop.*` columns instead of `family_names` and `given_names`.
+- Allow for generic `tags.*` columns instead of `difficulty` and `category` (multivalued, filterable).
 
 ## [1.1.0] - 2025-01-03